# docs/swagger.yaml
basePath: /v1
definitions:
  auth.confirmAccountRequest:
    properties:
      email:
        example: mike@protonmail.com
        type: string
    required:
    - email
    type: object
  auth.createNewPwdRequest:
    properties:
      guid:
        example: f47ac10b-58cc-8372-8567-0e02b2c3d479
        type: string
      password:
        example: secretControl
        maxLength: 30
        minLength: 8
        type: string
      token:
        example: eyJhbGciOiJIUzI1Ni
        type: string
    required:
    - guid
    - password
    - token
    type: object
  auth.loginRequest:
    properties:
      password:
        example: control123
        maxLength: 30
        minLength: 8
        type: string
      username:
        example: mrrobot or mr-robot@protonmail.com
        type: string
    required:
    - password
    - username
    type: object
  auth.resetPwdRequest:
    properties:
      email:
        example: mike@protonmail.com
        type: string
    required:
    - email
    type: object
  comment.CreateCommentRequest:
    properties:
      body:
        type: string
      sha256:
        type: string
      username:
        type: string
    required:
    - body
    - sha256
    type: object
  comment.UpdateCommentRequest:
    properties:
      body:
        type: string
    required:
    - body
    type: object
  entity.Comment:
    properties:
      body:
        description: Body represents the content of the comment.
        type: string
      id:
        description: ID represents the activity identifier.
        type: string
      sha256:
        description: |-
          SHA256 references the hash of the file
          where the comment has been made.
        type: string
      timestamp:
        description: Timestamp when this activity happened.
        type: integer
      type:
        description: Type represents the document type.
        type: string
      username:
        description: Username represents the author of the comment.
        type: string
    type: object
...
  errors.ErrorResponse:
    properties:
      details: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
host: api.saferwall.com
info:
  contact:
    email: support@saferwall.com
    name: API Support
    url: https://about.saferwall.com/contact.html
  description: Interact with Saferwall Malware Analysis Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://about.saferwall.com/tos
  title: Saferwall Web API
  version: "1.0"
paths:
  /auth/login/:
    post:
      consumes:
      - application/json
      description: Users logins by username and password.
      parameters:
      - description: Username and password
        in: body
        name: auth-request
        required: true
        schema:
          $ref: '#/definitions/auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "value"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Log in
      tags:
      - Authentication
...
  /auth/password/:
    post:
      consumes:
      - application/json
      description: Update the password from the auth token received in email.
      parameters:
      - description: New password request
        in: body
        name: reset-pwd
        required: true
        schema:
          $ref: '#/definitions/auth.createNewPwdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "value"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new password from a token received in email
      tags:
      - Authentication
  /auth/resend-confirmation/:
    post:
      consumes:
      - application/json
      description: Send a new confirmation email link to confirm user's account.
      parameters:
      - description: Account confirmation request
        in: body
        name: reset-pwd
        required: true
        schema:
          $ref: '#/definitions/auth.confirmAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "value"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Resend a confirmation email
      tags:
      - Authentication
  /auth/reset-password/:
    post:
      consumes:
      - application/json
      description: Request a reset password for anonymous users.
      parameters:
      - description: Email used during account sign-up
        in: body
        name: reset-pwd
        required: true
        schema:
          $ref: '#/definitions/auth.resetPwdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token": "value"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Reset password for non-logged users by email
      tags:
      - Authentication
  /auth/verify-account/:
    get:
      description: Verify the JWT token received during account creation.
      parameters:
      - description: GUID to identify the token
        in: query
        name: guid
        required: true
        type: string
      - description: JWT token generated for account creation
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"token": "value"}'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Confirm a new account creation
      tags:
      - Authentication
  /comments/:
    post:
      consumes:
      - application/json
      description: Create a new comment.
      parameters:
      - description: Comment body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create a new comment
      tags:
      - Comment
  /comments/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves information about a comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get comment by ID
      tags:
      - Comment
  /comments/{id}/:
    delete:
      description: Deletes a comment by ID.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/entity.Comment'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Deletes a comment
      tags:
      - Comment
    patch:
      consumes:
      - application/json
      description: Replace a cocument with a new comment's document.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: New comment data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/comment.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update a comment object (full update)
      tags:
      - Comment
...
schemes:
- https
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"